+++ R/list_user_data_packages.R
@@ -5,12 +5,15 @@
 #' distinguished name. Data packages that were uploaded by the specified user
 #' but have since been deleted are excluded from the list.
 #'
-#' @param dn (character) Distinguished name of user. Create with
-#' \code{create_dn()}.
+#' @param edi_id (character) The EDI-ID of the user. This ID can be obtained by 
+#' logging into the EDI Identity and Access Manager 
+#' (\url{https://auth.edirepository.org/auth/ui/signin}) and copying the 
+#' "EDI-ID" string from your profile home page.
 #' @param env (character) Repository environment. Can be: "production",
 #' "staging", or "development".
 #'
-#' @return (character) Data package identifiers belonging to a \code{dn}
+#' @return (character) Data package identifiers belonging to a 
+#'   \code{edi_id}
 #' 
 #' @family Listing
 #'
@@ -20,13 +23,13 @@
 #' \dontrun{
 #' 
 #' # List user data packages
-#' dn <- create_dn(userId = "dbjourneynorth")
-#' packageIds <- list_user_data_packages(dn)
+#' edi_id <- "EDI-543afa80c859825d35d37d9111c24a4a65a0ff9e"
+#' packageIds <- list_user_data_packages(edi_id = edi_id)
 #' packageIds
 #' #> [1] "edi.948.1" "edi.949.1"
 #' }
-list_user_data_packages <- function(dn, env = "production") {
-  url <- paste0(base_url(env), "/package/user/", dn)
+list_user_data_packages <- function(edi_id, env = "production") {
+  url <- paste0(base_url(env), "/package/user/", edi_id)
   resp <- httr::GET(url, set_user_agent(), handle = httr::handle(""))
   res <- httr::content(resp, as = "text", encoding = "UTF-8")
   httr::stop_for_status(resp, res)