+++ R/utilities.R
@@ -123,6 +123,10 @@ config_test_eml <- function(userId, url) {
 #' written to file
 #' @param packageId (character) Package identifier, of the form
 #' "scope.identifier.revision", for the new EML file
+#' @param edi_id (character) The EDI-ID of the user. This ID can be obtained by 
+#' logging into the EDI Identity and Access Manager 
+#' (\url{https://auth.edirepository.org/auth/ui/signin}) and copying the 
+#' "EDI-ID" string from your profile home page.
 #'
 #' @return (character) Full path to EML file written by this function to
 #' \code{path}. Should be \code{tempdir()} if executed in a testthat context.
@@ -132,16 +136,15 @@ config_test_eml <- function(userId, url) {
 #'
 #' @noRd
 #'
-create_test_eml <- function(path, packageId) {
+create_test_eml <- function(path, packageId, edi_id) {
   # Read EML template
   eml <- system.file("extdata", "eml.xml", package = "EDIutils")
   eml <- xml2::read_xml(eml)
   # Add packageId
   xml2::xml_attr(eml, "packageId") <- packageId
   # Add principal
-  dn <- create_dn(Sys.getenv("EDI_USERID"))
   principal <- xml2::xml_find_first(eml, ".//principal")
-  xml2::xml_text(principal) <- dn
+  xml2::xml_text(principal) <- edi_id
   # Add URL
   url <- xml2::xml_find_first(eml, ".//online/url")
   xml2::xml_text(url) <- Sys.getenv("EDI_TEST_URL")