+++ tests/testthat/test_evaluate_data_package.R
@@ -6,7 +6,8 @@ testthat::test_that("Test attributes of returned object", {
   # Create data package for evaluation
   identifier <- create_reservation(scope = "edi", env = "staging")
   packageId <- paste0("edi.", identifier, ".1")
-  eml <- create_test_eml(path = tempdir(), packageId = packageId)
+  eml <- create_test_eml(path = tempdir(), packageId = packageId, 
+                         edi_id = "EDI-543afa80c859825d35d37d9111c24a4a65a0db9f")
   on.exit(file.remove(eml), add = TRUE, after = FALSE)
   # Evaluate
   transaction <- evaluate_data_package(eml, env = "staging")