+++ NEWS.md
@@ -7,6 +7,7 @@ EDIutils 1.1.0 (2026-01-06)
 ### MINOR IMPROVEMENTS
 
   * Update authentication to support new EDI IAM system (#60)
+  * Update tests to handle new fields from various API methods
 
 EDIutils 1.0.3 (2023-10-10)
 ===========================